requirements.txt
altair==4.0.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
entrypoints==0.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==3.20.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.12.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.3
zipp==3.19.0
